Renamed a few packages to internal in the cometvisu bundle. See openhab/static-code-analysis#100
