+++ addons/ui/org.openhab.ui.cometvisu/OSGI-INF/cv-backend-charts.xml
@@ -12,7 +12,7 @@
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" enabled="true" immediate="true" name="org.openhab.ui.cometvisu.backend.ChartResource">
    <implementation class="org.openhab.ui.cometvisu.backend.ChartResource"/>
    <service>
-      <provide interface="org.openhab.ui.cometvisu.backend.ChartResource"/>
+      <provide interface="org.openhab.ui.cometvisu.internal.backend.ChartResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setItemRegistry" cardinality="1..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="static" unbind="unsetItemRegistry"/>
