+++ addons/ui/org.openhab.ui.cometvisu/OSGI-INF/cv-backend-config.xml
@@ -10,9 +10,9 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.ui.cometvisu.backend.ConfigResource">
-   <implementation class="org.openhab.ui.cometvisu.backend.ConfigResource"/>
+   <implementation class="org.openhab.ui.cometvisu.internal.backend.ConfigResource"/>
    <service>
-      <provide interface="org.openhab.ui.cometvisu.backend.ConfigResource"/>
+      <provide interface="org.openhab.ui.cometvisu.internal.backend.ConfigResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
 </scr:component>
