+++ addons/ui/org.openhab.ui.cometvisu/OSGI-INF/cv-backend-read.xml
@@ -10,9 +10,9 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.ui.cometvisu.backend.ReadResource">
-   <implementation class="org.openhab.ui.cometvisu.backend.ReadResource"/>
+   <implementation class="org.openhab.ui.cometvisu.internal.backend.ReadResource"/>
    <service>
-      <provide interface="org.openhab.ui.cometvisu.backend.EventBroadcaster"/>
+      <provide interface="org.openhab.ui.cometvisu.internal.backend.EventBroadcaster"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setItemRegistry" cardinality="1..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="static" unbind="unsetItemRegistry"/>
