+++ addons/ui/org.openhab.ui.cometvisu/OSGI-INF/cv-backend-write.xml
@@ -10,9 +10,9 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.ui.cometvisu.backend.WriteResource">
-   <implementation class="org.openhab.ui.cometvisu.backend.WriteResource"/>
+   <implementation class="org.openhab.ui.cometvisu.internal.backend.WriteResource"/>
    <service>
-      <provide interface="org.openhab.ui.cometvisu.backend.WriteResource"/>
+      <provide interface="org.openhab.ui.cometvisu.internal.backend.WriteResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setItemRegistry" cardinality="1..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="static" unbind="unsetItemRegistry"/>
