+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/ConfigResource.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.ui.cometvisu.backend;
+package org.openhab.ui.cometvisu.internal.backend;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
