+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/ReadResource.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.ui.cometvisu.backend;
+package org.openhab.ui.cometvisu.internal.backend;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -37,8 +37,8 @@ import org.eclipse.smarthome.io.rest.RESTResource;
 import org.glassfish.jersey.media.sse.EventOutput;
 import org.glassfish.jersey.media.sse.SseBroadcaster;
 import org.glassfish.jersey.media.sse.SseFeature;
-import org.openhab.ui.cometvisu.backend.beans.StateBean;
 import org.openhab.ui.cometvisu.internal.Config;
+import org.openhab.ui.cometvisu.internal.backend.beans.StateBean;
 import org.openhab.ui.cometvisu.internal.listeners.StateEventListener;
 import org.openhab.ui.cometvisu.internal.util.SseUtil;
 import org.slf4j.Logger;
