+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/beans/StateBean.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2010-2017 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.openhab.ui.cometvisu.internal.backend.beans;
+
+/**
+ * Item bean for broadcasted item states.
+ * 
+ * @author Tobias Bräutigam - Initial Contribution and API
+ * @since 2.0.0
+ */
+public class StateBean {
+
+    public String name;
+
+    public String state;
+
+}
