+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/backend/beans/SuccessBean.java
@@ -6,18 +6,16 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.ui.cometvisu.backend.beans;
+package org.openhab.ui.cometvisu.internal.backend.beans;
 
 /**
  * This is a java bean that is used with JAXB to define the login entry
  * page of the Cometvisu interface.
- *
+ * 
  * @author Tobias Bräutigam
  * @since 2.0.0
  *
  */
-public class LoginBean {
-    public String v;
-    public String s;
-    public ConfigBean c;
+public class SuccessBean {
+    public int success;
 }
