+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/config/ConfigHelper.java
@@ -61,7 +61,7 @@ import org.openhab.ui.cometvisu.internal.config.beans.Stylings;
 import org.openhab.ui.cometvisu.internal.config.beans.Text;
 import org.openhab.ui.cometvisu.internal.config.beans.Trigger;
 import org.openhab.ui.cometvisu.internal.config.beans.Widgetinfo;
-import org.openhab.ui.cometvisu.servlet.CometVisuApp;
+import org.openhab.ui.cometvisu.internal.servlet.CometVisuApp;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
