+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/config/VisuConfig.java
@@ -59,7 +59,7 @@ import org.openhab.ui.cometvisu.internal.config.beans.Text;
 import org.openhab.ui.cometvisu.internal.config.beans.Trigger;
 import org.openhab.ui.cometvisu.internal.config.beans.Video;
 import org.openhab.ui.cometvisu.internal.config.beans.Web;
-import org.openhab.ui.cometvisu.servlet.CometVisuApp;
+import org.openhab.ui.cometvisu.internal.servlet.CometVisuApp;
 import org.rrd4j.ConsolFun;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
