+++ addons/ui/org.openhab.ui.cometvisu/src/main/java/org/openhab/ui/cometvisu/internal/util/SseUtil.java
@@ -11,8 +11,8 @@ package org.openhab.ui.cometvisu.internal.util;
 import javax.ws.rs.core.MediaType;
 
 import org.glassfish.jersey.media.sse.OutboundEvent;
-import org.openhab.ui.cometvisu.backend.beans.StateBean;
 import org.openhab.ui.cometvisu.internal.StateBeanMessageBodyWriter;
+import org.openhab.ui.cometvisu.internal.backend.beans.StateBean;
 
 /**
  * Utility class containing helper methods for the SSE implementation.
